lambda_functions/login_manager/requirements.txt
pyjwt
argon2-cffi